+++ commentary/abstracts.py
@@ -26,7 +26,7 @@ class BaseCommentAbstractModel(models.Model):
         ContentType, verbose_name=_('content type'),
         related_name='content_type_set_for_%(class)s', on_delete=models.CASCADE
     )
-    object_pk = models.CharField(_('object ID'), max_length=255)
+    object_pk = models.CharField(_('object ID'), max_length=64, db_index=True)
     content_object = GenericForeignKey(
         ct_field='content_type', fk_field='object_pk'
     )
@@ -43,7 +43,7 @@ class BaseCommentAbstractModel(models.Model):
         )
     )
     is_removed = models.BooleanField(
-        _('is removed'), default=False, help_text=_(
+        _('is removed'), default=False, db_index=True, help_text=_(
             'Check this box if the comment is inappropriate. A "This '
             'comment has been removed" message will be displayed instead.'
         )